Move RPC modules to application.
Tidy as per feedback in https://github.com/tauri-apps/wry/pull/83.
