+++ src/platform/mod.rs
@@ -27,8 +27,6 @@ use std::{collections::HashMap, sync::Mutex};
 use once_cell::sync::Lazy;
 use serde_json::Value;
 
-pub(crate) const RPC_CALLBACK_NAME: &str = "__rpc__";
-
 pub(crate) static CALLBACKS: Lazy<
     Mutex<
         HashMap<
@@ -52,54 +50,4 @@ struct RPC {
     params: Vec<Value>,
 }
 
-/// Function call from Javascript.
-///
-/// If the callback name matches the name for an RPC handler
-/// the payload should be passed to the handler transparently.
-///
-/// Otherwise attempt to find a `Callback` with the same name
-/// and pass it the payload `params`.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct FuncCall {
-    callback: String,
-    payload: RpcRequest,
-}
 
-/// RPC request message.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct RpcRequest {
-    jsonrpc: String,
-    pub id: Option<Value>,
-    pub method: String,
-    pub params: Option<Value>,
-}
-
-/// RPC response message.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct RpcResponse {
-    jsonrpc: String,
-    id: Option<Value>,
-    result: Option<Value>,
-    error: Option<Value>,
-}
-
-impl RpcResponse {
-
-    /// Create a new result response.
-    pub fn new_result(id: Option<Value>, result: Option<Value>) -> Self {
-        Self {
-            jsonrpc: "2.0".to_string(),
-            id, result,
-            error: None
-        } 
-    }
-
-    /// Create a new error response.
-    pub fn new_error(id: Option<Value>, error: Option<Value>) -> Self {
-        Self {
-            jsonrpc: "2.0".to_string(),
-            id, error,
-            result: None
-        } 
-    }
-}
